+++ Controllers/OldTagController.cs
@@ -8,20 +8,20 @@ using Microsoft.EntityFrameworkCore;
 using DatabaseAPI.Model;
 
 namespace DatabaseAPI.Controllers {
-    [Route("auth0-tags")]
+    [Route("old/tags")]
     [ApiController]
-    public class Auth0TagController : ControllerBase {
+    public class OldTagController : ControllerBase {
         private readonly DHBWExpertsdatabaseContext _context;
 
         //The context is managed by the WEBAPI and used here via Dependency Injection.
-        public Auth0TagController(DHBWExpertsdatabaseContext context) {
+        public OldTagController(DHBWExpertsdatabaseContext context) {
             _context = context;
         }
 
         // GET: /Users/contacts/5
         //The user assosiated contacts of the user a returned
-        [HttpGet("{id:int}/validations", Name = "getValidationsByTagId")]
-        public async Task<ActionResult<IEnumerable<Object>>> getValidationsByTagId(int id) {
+        [HttpGet("{id:int}/validations", Name = "getValidationsByTagIdOld")]
+        public async Task<ActionResult<IEnumerable<Object>>> getValidationsByTagIdOld(int id) {
 
             var query =
                 from val in _context.TagValidation
@@ -44,17 +44,17 @@ namespace DatabaseAPI.Controllers {
 
         // GET: /Users/contacts/5
         //The user assosiated contacts of the user a returned
-        [HttpGet("{id:int}", Name = "getTagByTagId")]
-        public async Task<ActionResult<IEnumerable<Object>>> getTagByTagId(int id) {
+        [HttpGet("{id:int}", Name = "getTagByTagIdOld")]
+        public async Task<ActionResult<IEnumerable<Object>>> getTagByTagIdOld(int id) {
 
             var query =
-                from tags in _context.Auth0Tags
+                from tags in _context.Tag
                 where tags.TagId == id
                 select new {
                     tagId = tags.TagId,
-                    tag = tags.Tag,
+                    tag = tags.Tag1,
                     user = tags.User,
-                    createdAt = tags.CreatedAt
+                    tmsCreated = tags.TmsCreated
                 };
 
             var result = await query.ToListAsync();
@@ -63,6 +63,29 @@ namespace DatabaseAPI.Controllers {
                 return NotFound();
             }
             return result;
+        }
+
+
+        // GET: /Users/contacts/5
+        //The user assosiated contacts of the user a returned
+        [HttpDelete("{id:int}", Name = "deleteTagByTagIdOld")]
+        public async Task<ActionResult> deleteTagByTagIdOld(int id) {
+
+            var tag = await _context.Tag.FindAsync(id);
+
+            if (tag is null) {
+                return NotFound();
+            }
+
+            _context.Tag.Remove(tag);
+
+            try {
+                await _context.SaveChangesAsync();
+            } catch (DbUpdateException) {
+                return Conflict();
+            }
+
+            return Ok();
         }
 
     }
