+++ Controllers/TagValidationController.cs
@@ -20,18 +20,18 @@ namespace DatabaseAPI.Controllers {
 
         // GET: /Users/contacts/5
         //The user assosiated contacts of the user a returned
-        [HttpGet("{id:int}", Name = "getValidationByValIdOld")]
-        public async Task<ActionResult<Object>> getValidationsByTagIdOld(int id) {
+        [HttpGet("{id:int}", Name = "getValidationByValId")]
+        public async Task<ActionResult<Object>> getValidationsByTagId(int id) {
 
             var query =
-                from val in _context.TagValidation
+                from val in _context.Auth0TagValidations
                 where val.ValidationId == id
                 select new {
                     validationId = val.ValidationId,
                     tag = val.Tag,
                     validatedBy = val.ValidatedBy,
                     comment = val.Comment,
-                    tmsCreated = val.TmsCreated
+                    CreatedAt = val.CreatedAt
                 };
 
             var result = await query.FirstOrDefaultAsync();
@@ -44,16 +44,16 @@ namespace DatabaseAPI.Controllers {
 
         // GET: /Users/contacts/5
         //The user assosiated contacts of the user a returned
-        [HttpDelete("{id:int}", Name = "deleteTagValidationByTagIdOld")]
-        public async Task<ActionResult> deleteTagValByValIdOld(int id) {
+        [HttpDelete("{id:int}", Name = "deleteTagValidationByTagId")]
+        public async Task<ActionResult> deleteTagValByValId(int id) {
 
-            var val = await _context.TagValidation.FindAsync(id);
+            var val = await _context.Auth0TagValidations.FindAsync(id);
 
             if (val is null) {
                 return NotFound();
             }
 
-            _context.TagValidation.Remove(val);
+            _context.Auth0TagValidations.Remove(val);
 
             try {
                 await _context.SaveChangesAsync();
